feat: Refactor CapacityComparisonGraph component to use static data
